RC6.5
Rifles now have 65% less ammo.
Fixes guns not being enchantable in survival, but you need to use an anvil.
New Material: HoCuFe[Hoe-Kyu-Fey] made of iron and copper, with a bitta bee wax.
New Material: Nurukukan Ingot: Alloy of Voidmetal and HoCuFe!
Textures are... ehhhhh for both.
New Trait: Eternal! Was that 6 shots or 7? 5% chance per level to not consume ammo on fire OR RETURN ammo to the stack when reloading!